requirements.txt
argparse==1.2.1
beautifulsoup4==4.4.1
bs4==0.0.1
feedparser==5.2.1
future==0.15.2
python-telegram-bot==4.2.0
wsgiref==0.1.2
